source/pythonScript/requirements.txt
boto3==1.21.6
botocore==1.24.6
cachetools==5.0.0
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
distro==1.7.0
et-xmlfile==1.1.0
ghostscript==0.7
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
jmespath==0.10.0
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
openpyxl==3.0.9
pandas==1.4.1
pdf2image==1.16.0
pdfminer.six==20211012
pdftopng==0.2.3
Pillow==9.0.1
poppler-utils==0.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.1
six==1.16.0
tabula-py==2.3.0
tabulate==0.8.9
textract-trp==0.1.3
tk==0.1.0
uritemplate==4.1.1
urllib3==1.26.8
